uz to vypisuje

+++ Core/HTS221/HTS221.h
@@ -22,8 +22,10 @@
 
 // Inicializácia senzora
 uint8_t HTS221_WriteRegister(uint8_t reg_address, uint8_t data);
-uint8_t HTS221_Init(void (*i2c_read)(uint8_t, uint8_t, uint8_t*, uint16_t),
-                    void (*i2c_write)(uint8_t, uint8_t, uint8_t*, uint16_t));
+uint8_t HTS221_Init(void (*i2c_read)(uint16_t, uint16_t, uint8_t*, uint16_t),
+                    void (*i2c_write)(uint16_t, uint16_t, uint8_t*, uint16_t));
+//uint8_t HTS221_Init(HAL_StatusTypeDef (*i2c_read)(uint16_t, uint16_t, uint8_t*, uint16_t),
+//					HAL_StatusTypeDef (*i2c_write)(uint16_t, uint16_t, uint8_t*, uint16_t));
 
 // Funkcie na čítanie hodnôt
 float HTS221_ReadTemperature(void);
